replace old gallery component with new one
